moved modal logic to courseSearch, removed some comments that don't matter anymore
